+++ src/components/content/editForm/Content.edit.display.js
@@ -1,5 +1,14 @@
 export default [
   {
+    type: 'textfield',
+    input: true,
+    key: 'className',
+    weight: 60,
+    label: 'CSS Class',
+    placeholder: 'CSS Class',
+    tooltip: 'The CSS class for this HTML element.'
+  },
+  {
     weight: 700,
     type: 'checkbox',
     label: 'Refresh On Change',